Exceptions fixed, tables refactor, delete on cascade

+++ demo/src/main/java/com/example/demo/services/implementation/WorkerServiceImp.java
@@ -8,11 +8,14 @@ import com.example.demo.exceptions.workerExceptions.WorkerConflictException;
 import com.example.demo.exceptions.workerExceptions.WorkerNotFoundException;
 import com.example.demo.exceptions.workerExceptions.WorkerUnauthorizedException;
 import com.example.demo.mapper.WorkerMapper;
+import com.example.demo.repositories.PayrollRepository;
+import com.example.demo.repositories.TransferRepository;
 import com.example.demo.repositories.WorkerRepository;
 import com.example.demo.services.interfaces.WorkerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -24,36 +27,43 @@ public class WorkerServiceImp implements WorkerService {
     private String LORE;
     @Autowired
     private WorkerRepository workerRepository;
+    @Autowired
+    private PayrollRepository payrollRepository;
+    @Autowired
+    private TransferRepository transferRepository;
     //endregion
 
     //region PUBLIC_METHODS
     @Override
     public WorkerDTO saveWorker(Worker worker) throws WorkerConflictException, WorkerBadRequestException {
         Worker workerDb = getWorker(worker.getDni());
-        if (workerDb != null) throw new WorkerConflictException("The worker already exist");
+        if (workerDb != null) throw new WorkerConflictException();
 
         String dni = worker.getDni();
         if (dni.length() != 9 || !Character.isLetter(dni.charAt(dni.length() - 1)))
-            throw new WorkerBadRequestException("Dni bad formated");
+            throw new WorkerBadRequestException();
         String name = worker.getName();
-        if (name.isEmpty() || name.isBlank())
-            throw new WorkerBadRequestException("The name is not correct");
+        if (name.isEmpty() || name.isBlank()) throw new WorkerBadRequestException();
 
         return WorkerMapper.INSTANCE.modelToDTO(workerRepository.save(worker));
     }
 
     @Override
+    @Transactional
     public void deleteWorker(String id) {
+        if (id == null) return;
         Worker fromDB = getWorker(id);
         if (fromDB == null) return;
+        deleteCascade(fromDB);
+
         workerRepository.delete(fromDB);
     }
 
     @Override
     public WorkerDTO riseSalary(String id, double amount) throws WorkerNotFoundException, TransferBadRequestException {
         Worker fromDB = getWorker(id);
-        if (fromDB == null) throw new WorkerNotFoundException("The worker is not found");
-        if (amount < 0) throw new TransferBadRequestException("The amount must be a valid number");
+        if (fromDB == null) throw new WorkerNotFoundException();
+        if (amount < 0) throw new TransferBadRequestException();
         fromDB.setSalary(fromDB.getSalary() + amount);
         workerRepository.save(fromDB);
         return WorkerMapper.INSTANCE.modelToDTO(fromDB);
@@ -62,12 +72,12 @@ public class WorkerServiceImp implements WorkerService {
     @Override
     public WorkerDTO workerInformation(String id) throws WorkerNotFoundException {
         Worker result = getWorker(id);
-        if (result == null) throw new WorkerNotFoundException("The provided worker does not exist");
+        if (result == null) throw new WorkerNotFoundException();
         return WorkerMapper.INSTANCE.modelToDTO(result);
     }
 
     public List<WorkerDTO> getAllWorkers(String pass) throws WorkerUnauthorizedException {
-        if (!LORE.equals(pass)) throw new WorkerUnauthorizedException("Unauthorized");
+        if (!LORE.equals(pass)) throw new WorkerUnauthorizedException();
         return workerRepository.findAll().stream().map(WorkerMapper.INSTANCE::modelToDTO).collect(Collectors.toList());
     }
     //endregion
@@ -75,6 +85,12 @@ public class WorkerServiceImp implements WorkerService {
     //region PRIVATE_METHODS
     private Worker getWorker(String id) {
         return workerRepository.findByDni(id);
+    }
+
+    private void deleteCascade(Worker worker) {
+        worker.getPayrolls().forEach(payrollId -> payrollRepository.deleteById((payrollId)));
+        worker.getTransfersEmitted().forEach(transferId -> transferRepository.deleteById(transferId));
+        worker.getTransfersReceived().forEach(transferId -> transferRepository.deleteById(transferId));
     }
     //endregion
 }
